requirements.txt
ipykernel
google-genai
pillow
fastapi
uvicorn
python-dotenv